I implemented a system to automatically generate all resolutions, although support for loading from the resolutions.cfg file can also be added back in.
I also went halfway through implementing different resolutions for fullscreen and widescreen that the game will remember.  But it really doesn't seem to work well at all, and I don't know what to do to fix it.  Gerstrong, could you please finish this up?  Then when you switch between fullscreen and windowed mode it will have two different resolutions that you normally have instead of having to change it when you do fullscreen and vice versa.

git-svn-id: https://clonekeenplus.svn.sourceforge.net/svnroot/clonekeenplus/cgenius/trunk@420 4df4b0f3-56ce-47cb-b001-ed939b7d65a6

+++ src/sdl/COpenGL.h
@@ -22,7 +22,7 @@ public:
 	virtual ~COpenGL();
 
 	void render(void);
-	bool initGL(unsigned Width, unsigned Height, unsigned char Depth,
+	bool initGL(float Width, float Height, unsigned char Depth,
 				GLint oglfilter, unsigned char scalex, bool aspect);
 
 	void setSurface(SDL_Surface *blitsurface);
